requirements.txt
esper
pre-commit
pygame-ce
pygbag
pyinstaller
